Fix usage of copyfiles
The last argument to `copyfiles` is always interpreted as a directory,
which is created if it does not exist. This means that previously,
`preact.d.ts` and `preact.js.flow` directories were being made in
`dist`.

This commit fixes the usage of `copyfiles` to copy the files to
the `dist` directory directly, and point to them in package.json
where appropriate.