Fix duplicate property with calc + variables
Reason is Safari 10 which doesn't parse css
variables inside of calc attributes correctly.
For properties that had a -webkit prefix back
in its day it will add both the unprefixed and
the prefixed version to the dom. But whenever
the value consists of calc() and a variable
it gets confused and applies the unprefixed
variant twice...