(feat) - debugging effect hooks (#1495)
This debugs the passing of an arguments array to hooks that need it. This avoids nasty endless calling bugs. This throws an error to breaking that endless calling.

Or is this actually a pattern that people want calling an effect on every render, personally it always bites me...

This could potentially be extended to useMemo and useCallback, warning about not providing an arguments array makes it a useless function call.

Warn about hooks in classes, I don't know if this is possible?