Add functions to `debug` for inspecting the rendered VNode tree
These are initially intended to be used by the Enzyme adapter, but they may
be useful in future for other development/testing tools needing to
access the rendered tree.

The API only exposes the minimal amount of information that the Enzyme
adpater needs, with the hope of not getting in the way of future changes
to Preact's internals.