Ignore code coverage reporting for devtools.
The devtools code uses newer ECMAScript features because the
devtools extension is only available for Firefox and Chrome. But
we didn't have such a distinction in our tests suite. That leads
to failing tests on IE11.

We tried running those tests only on FF and Chrome, but because
karma-coverage always collects coverage data from the latest
running browser, it would lead to coverage failures. The latest
browser is typically slowest and you may have guessed it already:
This will pretty much always be IE11.

The nearterm conclusion is to just disable coverage reporting
for the devtools code completely...