Fix useContext args change check (+9 B)
The previous change to destructure the args of getHookState broke because `state._value` was no longer modified. Instead the local `_value` variable was modified.

That change saved bytes because terser recognized that a local variable was modified and never read so it removed the variable entirely.

Added a test to capture the desired behavior