Catch exceptions when flushing effects during `act`
Make sure that future state updates and effects continue to run if an
exception occurs during a call to `act`.

This required resolving an issue with hooks that an exception inside
`useEffect` left the effect callback in the effect queue, so it would
run again the next time effects were flushed.