Simplify diffChildren patch vs mount logic
This commit breaks suspense-hydration. It reveals that we call
`.forceUpdate` on component that suspened during hydration. In our
current diff this is invalid because we recreate and remount components
that suspend during hydration so the component that Suspense has a
reference to is no longer valid and calling `.forceUpdate` on it
triggers a remount of the component.

This previously worked because renderComponent calls `diff` which
resumes hydration and diffChildren called mount with the previous
hydration state it was passed even if there wasn't an oldVNode.
This change changes how diffChildren invokes mount such that it
doesn't pass down the hydration state if there is no oldVNode. This
change is the right direction because eventually diffChildren will no
longer have params to receive hydration state to pass into mount.