Combine mountComponent and patchComponent into renderComponent
Also, remove the old diff/index.js.

Re: the change in the suspense - Previously Suspense would initial render
something like:
```jsx
<Fragment> <!-- Fragment 1 -->
  <Lazy />
</Fragment>
<Fragment> <!-- Fragment 2 -->
  {null}
</Fragment>
```
and when the Suspense component suspends, it renders
```jsx
<Fragment>
  {null}
</Fragment>
<Fragment>
  {fallback}
</Fragment>
```

Also when Suspense suspends, the `_component` property of the oldVNodes that
suspended (Fragment 1 and Lazy above) is nullified. This behavior means that
when the Fragment 1 diffed after a suspend, it will have an oldVNode but no
_component property (a weird behavior...). Seeing a `null` _component property,
the diff happily created one for it and went about diffing the rest of the its
children. In this case the old children would be `Lazy` and the new children
`null`. The DOM associated with `Lazy` would be unmounted but since it's
`_component` property is null, no lifcycles will be called (as expected).

With this change however, mounting a component's children and patching a
component's children are two different code paths. So when Fragment 1 comes in
without a `_component` property, the "mounting children" code path is taken.
However, the mounting code path doesn't do any unmounting. So the DOM of the
Lazy component is never unmounted (which is a bug).

This commit changes Suspense so that when it suspends, it renders
```jsx
{null}
<Fragment>
  {fallback}
</Fragment>
```

Now when the children of Suspense are diffed, the `null` is compared against
Fragment 1 and so Fragment 1 and everything under Fragment 1 is unmounted (as
expected). And still no lifcycles are are invoked since `_component` property is
still nullified (as expected). This behavior means that no VNodes with nulled
`_component` properties enter our diff.