Don't modify internal._children in place
As we patch children, we mark children as used by setting their value in the children array to `undefined`. If a child is never used, it remains in the children array and we then unmount at the end.

This behavior means we modify the children array of the internal as we diff, and only after we successfully diffed the all children, do we swap the new, correct children array for the holey oldChildren array.

The new behavior is to copy the oldChildren array and modify this copy, leaving the internal._children array intact. At this point we still build a new children array and replace the old children on the internal at the end.

At some future point, we should consider replacing the children in `internal._children` as we diff if that improves perf or memory. I didn't do that this time cuz it breaks some more tests.

This PR fixes the remaining 3 suspense tests.