WIP: Experiement not allocating children arrays for single children
Currently mountChildren and mountChild are two different functions but diffChildren still coerces the single child into an array but unboxes it when setting it on the _children property.

This doesn't perform as well as I'd hoped on 02_replace1k benchmark as it incurs more nonconcurrent optimizations on the main thread (5) instead of the previous version which only has 2.