Reorder children when promise is thrown and queue cDU only on successful render
When we are re-rooting a suspenses children, we need to re-parent the children of the component that threw, so call `reorderChildren` when a thenable is thrown.

Also, we shouldn't queue componentDidUpdate everytime we attempt to render a component (it could throw). Instead only queue cDU when the component successfully renders.

Adjust some of the tests to the new Suspense behavior. Since we no longer set force on every component under Suspense, some of the intermediate components don't rerender upon resuming. This new behavior actually mirrors the behavior that existed in our original suspense implementation and was changed in ea8521f05cd1128b8222d306a774d08be974b726 .