Add test for mounting a Suspense component after a render
Re: the removed TODO: This situation doesn't happen because in the render where Lazy throws, it's siblings and parent finish rendering as if nothing happened. So all the parent pointers above and around Lazy are set up correctly. So when the fallback goes to render and getDomSibling is called for the Suspense portal, parent pointers have the correct children (they are the same children as what was just rendered and suspended) so getDomSibling works correctly.