Revert "Memoize entity encoding"
This reverts commit 1bc5d56b839642204598f9f21b420587a52b8149.

Because we cannot forsee the length and amount of strings we need to
be more careful about keeping them around as memory consumption
can grow unbounded. In the future we should check if we can use
a limited cache, only cache attributes or something else. WeakMaps
unfortunately won't help as they cannot be used with primitives as
cache key.