Fix option focus bug caused by IE11 option value workaround.
Whenever children of an <option> element are changed Preact
resets the value to an empty string. This is a hack for an IE11
bug triggered whenever option value field is inferred from the
text inside the element.

This hack breaks the behaviour of <select> element on other browsers:
if you click on the dropdown, mouse over another option (but don't
click) and rendering happens at this moment, the focus will go back
to the current element. This is caused by setting the value attribute
on <option> elements even though it isn't necessary (the value didn't
change since the last time but Preact forces it to be set anyway).

To resolve the issue, we save the value dom attribute before we place
children and then check if it has changed afterwards, as a way to detect
whether we need to activate the hack.