Add globally shared rendererState variable
This commit introduces a rendererState shared "global" variable that contains the state of the renderer we want to share with a component implementation. Currently, it just holds the current context for this component's subtree. Because the globalContext is now managed... well.. globally... we no longer need to pass it down as a parameter to all the functions.

This commit also changes has context is managed. Instead of every component storing its globalContext on itself, only the components that change the context will store a reference to it, and this reference is now stored on Internals. So when we rerender an internal, we climb up the tree to the nearest internal that modified context, stopping at the root, which has context set to an empty object.