fix TargetedEvent target definition
The Typescript definitions for events are still painful to use, as there are many conflicts with existing UI libraries. For example, @mui/material is built around the .target property. The typscript definitions of preact will cause the developer a headache as the typscript definition for the ChangeEvent does not work with .target or .currentTarget without a few lines of Typscript pain around it.