Bring back _renderCallbacks on component
Bring back _renderCallbacks on component to fix Suspense test. I believe this fixes the test because it allows queued render callbacks (namely componentDidMount) to persist across suspensions. So cDM is queued on initial render when the component suspends. When the component resumes, cDM is still in the queue and is flushed and called on resumption.