fix: prevent invalid DOM nesting false positives
When Preact is rendering into a container in the middle of the page we
attempted to walk outside the render root via the DOM tree. Problem is
that that doesn't work on the server where there is no DOM, which
lead to us spamming the terminal with false positive warnings.

So this skips the DOM validation if we reach the render root.